requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.5.0
pyyaml>=6.0.1
mkdocs>=1.5.0
pytest>=7.4.0
httpx>=0.27.0

# 数据库和序列化依赖
SQLAlchemy>=2.0.43
orjson>=3.11.3
requests>=2.32.5
requests-toolbelt>=1.0.0
urllib3>=2.5.0

# 数据压缩和文本处理
zstandard>=0.25.0
charset-normalizer>=3.4.3
greenlet>=3.2.4

# AI模型相关依赖
openai>=1.0.0
zhipuai>=2.0.0
#azure-openai>=1.0.0
dashscope>=1.14.0

# LangChain框架依赖
langchain>=0.3.27
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.3.76
langchain-text-splitters>=0.3.11

# LangChain相关工具和支持包
langsmith>=0.4.31
tenacity>=9.1.2
jsonpatch>=1.33
jsonpointer>=3.0.0

